+++ tailwind.config.js
@@ -10,6 +10,8 @@ module.exports = {
   },
 
   screens: {
+    'xs': '320px',
+    'sm': '640px',
     'md': '768px',
     tablet: '960px',
     'xl': '1280px',
